chore(wfs): simplify proxy handling (refs #13)
