+++ package.json
@@ -39,7 +39,6 @@
     "commander": "^14.0.3",
     "js-yaml": "^4.1.1",
     "minisearch": "^7.2.0",
-    "proxy-from-env": "^2.1.0",
     "undici": "^6.23.0",
     "zod": "^4.3.6"
   },
